Trying out Anca's suggestion of separating the Xgboost 4j and Predictor implementations in separate sbt subprojects
