+++ mleap-xgboost-predictor/build.sbt
@@ -1,0 +1,4 @@
+import ml.combust.mleap.{Dependencies, Common}
+
+Common.defaultMleapSettings
+Dependencies.xgboostPredictorRuntime
