+++ mleap-xgboost-predictor/src/main/resources/reference.conf
@@ -1,0 +1,5 @@
+ml.combust.mleap.xgboost.ops = [
+  "ml.combust.mleap.xgboost.predictor.bundle.ops.XGBoostPredictorClassificationOp",
+]
+
+ml.combust.mleap.registry.default.ops += "ml.combust.mleap.xgboost.ops"
