+++ mleap-xgboost-predictor/src/main/scala/ml/combust/mleap/xgboost/predictor/XGBoostPredictorClassification.scala
@@ -1,18 +1,18 @@
-package ml.combust.mleap.xgboost.runtime
+package ml.combust.mleap.xgboost.predictor
 
 import biz.k11i.xgboost.util.FVec
 import ml.combust.mleap.core.types.NodeShape
+import ml.combust.mleap.core.util.VectorConverters._
 import ml.combust.mleap.runtime.function.UserDefinedFunction
 import ml.combust.mleap.tensor.Tensor
 import ml.combust.mleap.runtime.frame.{MultiTransformer, Row, Transformer}
-import ml.combust.mleap.core.util.VectorConverters._
 import XgbConverters._
 
 
 case class XGBoostPredictorClassification(
-                                  override val uid: String = Transformer.uniqueName("xgboost.classification"),
-                                 override val shape: NodeShape,
-                                 override val model: XGBoostPredictorClassificationModel) extends MultiTransformer {
+                                           override val uid: String = Transformer.uniqueName("xgboost.classification"),
+                                           override val shape: NodeShape,
+                                           override val model: XGBoostPredictorClassificationModel) extends MultiTransformer {
   override val exec: UserDefinedFunction = {
 
     // Since the Predictor is our performant implementation, we only compute probability for performance reasons.
