+++ mleap-xgboost-predictor/src/main/scala/ml/combust/mleap/xgboost/predictor/struct/FVecFactory.scala
@@ -1,4 +1,4 @@
-package ml.combust.mleap.xgboost.runtime.struct
+package ml.combust.mleap.xgboost.predictor.struct
 
 import java.{lang, util}
