+++ mleap-xgboost-runtime/src/main/scala/ml/combust/mleap/xgboost/runtime/XgbConverters.scala
@@ -1,8 +1,6 @@
 package ml.combust.mleap.xgboost.runtime
 
-import biz.k11i.xgboost.util.FVec
 import ml.combust.mleap.tensor.{DenseTensor, SparseTensor, Tensor}
-import ml.combust.mleap.xgboost.runtime.struct.FVecFactory
 import ml.dmlc.xgboost4j.LabeledPoint
 import ml.dmlc.xgboost4j.scala.DMatrix
 import org.apache.spark.ml.linalg.{DenseVector, SparseVector, Vector}
@@ -19,15 +17,6 @@ trait XgbConverters {
           new DMatrix(Iterator(new LabeledPoint(0.0f, null, values.map(_.toFloat))))
       }
     }
-
-    def asXGBPredictor: FVec = {
-      vector match {
-        case sparseVector: SparseVector =>
-          FVecFactory.fromSparseVector(sparseVector)
-        case denseVector: DenseVector =>
-          FVecFactory.fromDenseVector(denseVector)
-      }
-    }
   }
 
   implicit class DoubleTensorOps(tensor: Tensor[Double]) {
@@ -38,16 +27,6 @@ trait XgbConverters {
 
         case DenseTensor(_, _) =>
           new DMatrix(Iterator(new LabeledPoint(0.0f, null, tensor.toDense.rawValues.map(_.toFloat))))
-      }
-    }
-
-    def asXGBPredictor: FVec = {
-      tensor match {
-        case sparseTensor: SparseTensor[Double] =>
-          FVecFactory.fromSparseTensor(sparseTensor)
-
-        case denseTensor: DenseTensor[Double] =>
-          FVecFactory.fromDenseTensor(denseTensor)
       }
     }
   }
