+++ mleap-xgboost-runtime/src/main/scala/ml/combust/mleap/xgboost/runtime/bundle/ops/XGBoostRegressionOp.scala
@@ -8,9 +8,10 @@ import ml.combust.bundle.dsl.{Model, Value}
 import ml.combust.bundle.op.OpModel
 import ml.combust.mleap.bundle.ops.MleapOp
 import ml.combust.mleap.runtime.MleapContext
-import ml.combust.mleap.xgboost.runtime.{XGBoostRegression, XGBoostRegressionModel}
+import ml.combust.mleap.xgboost.runtime._
 import ml.dmlc.xgboost4j.scala.XGBoost
 import resource._
+
 
 /**
   * Created by hollinwilkins on 9/16/17.
