+++ mleap-xgboost-runtime/src/test/scala/ml/combust/mleap/xgboost/runtime/testing/BundleSerializationUtils.scala
@@ -1,12 +1,13 @@
 package ml.combust.mleap.xgboost.runtime.testing
 
 import java.io.File
+import java.nio.file.Files
+import java.nio.file.Paths
 
 import ml.combust.bundle.BundleFile
 import ml.combust.bundle.serializer.SerializationFormat
 import ml.combust.mleap.runtime.{MleapContext, frame}
 import ml.combust.mleap.runtime.frame.Transformer
-import ml.combust.mleap.xgboost.runtime.bundle.ops.{XGBoostClassificationOp, XGBoostPredictorClassificationOp}
 import resource.managed
 
 
@@ -15,8 +16,9 @@ trait BundleSerializationUtils {
   def serializeModelToMleapBundle(transformer: Transformer): File = {
     import ml.combust.mleap.runtime.MleapSupport._
 
-    new File("/tmp/mleap/xgboost-runtime-parity").mkdirs()
-    val file = new File(s"/tmp/mleap/xgboost-runtime-parity/${this.getClass.getName}.zip")
+    val temporaryDir = Files.createTempDirectory("xgboost-runtime-parity").toString
+    val path = Paths.get(temporaryDir, s"${this.getClass.getName}.zip")
+    val file = new File(path.toUri)
     file.delete()
 
     for(bf <- managed(BundleFile(file))) {
@@ -35,14 +37,4 @@ trait BundleSerializationUtils {
     }).tried.get
   }
 
-  def loadXGBoostPredictorFromBundle(bundleFile: File)
-                                    (implicit context: MleapContext): frame.Transformer = {
-
-    // Register a different Op to change the deserialization class between tests.
-    // Use to deserialize with Predictor rather than xgboost4j
-    context.bundleRegistry.register(new XGBoostPredictorClassificationOp())
-    val transformer = loadMleapTransformerFromBundle(bundleFile)
-    context.bundleRegistry.register(new XGBoostClassificationOp())  // revert to the original Op
-    transformer
-  }
 }
