+++ project/Dependencies.scala
@@ -117,7 +117,9 @@ object Dependencies {
 
   val tensorflow = l ++= tensorflowDeps ++ Seq(Test.scalaTest)
 
-  val xgboostRuntime = l ++= Seq(xgboostDep) ++ Seq(xgboostPredictorDep) ++ Test.spark ++ Seq(Test.scalaTest)
+  val xgboostRuntime = l ++= Seq(xgboostPredictorDep) ++ Seq(xgboostDep) ++ Test.spark ++ Seq(Test.scalaTest)
+
+  val xgboostPredictorRuntime = l ++= Seq(xgboostPredictorDep) ++ Test.spark ++ Seq(Test.scalaTest)
 
   val xgboostSpark = l ++= Seq(xgboostSparkDep) ++ Provided.spark
